server/requirements.txt
fastapi
uvicorn
pillow
requests
python-multipart
google-generativeai
google-cloud-vision
